KAN-130/three-card-list-components-to-become-one (#94)
* refactor: extract card filtering and sorting logic into card_filter_service

* refactor: create layout strategy abstraction for card list management

* refactor: create render strategy abstraction for card list rendering

* refactor: introduce UnifiedViewStrategy to compose layout and render strategies

* refactor: simplify render_tasks to use unified view strategy

* refactor: update app initialization to use UnifiedViewStrategy

* refactor: simplify card handlers to work with unified view strategy

* refactor: simplify navigation handlers to work with unified view strategy

* chore: add changeset

* refactor: fix clippy warnings in handlers and render_strategy

- Change and_then(|x| Some(y)) to map(|x| y) in navigation_handlers
- Collapse else-if block in render_strategy for cleaner control flow

* refactor: extract common filtering logic in card_filter_service

Extract filter logic into apply_card_filters() helper function to eliminate duplication between filter_and_sort_cards() and filter_and_sort_cards_by_column(). Reduces duplicated code and improves maintainability.

* refactor: add as_any() to LayoutStrategy trait for symmetry

Add as_any() method alongside existing as_any_mut() for consistent trait design. Implement in both SingleListLayout and ColumnListsLayout. Improves future extensibility for immutable downcasting.

* refactor: simplify downcast paths by importing UnifiedViewStrategy

Add UnifiedViewStrategy imports to ui.rs and navigation_handlers.rs. Replace verbose crate::view_strategy::UnifiedViewStrategy paths with imported UnifiedViewStrategy for improved readability.

* chore: cargo fmt
