+++ .changeset/kan-130-three-card-list-components-to-become-one.md
@@ -1,0 +1,18 @@
+---
+bump: patch
+---
+
+A refactoring.
+
+- refactor: simplify navigation handlers to work with unified view strategy
+- refactor: simplify card handlers to work with unified view strategy
+- refactor: update app initialization to use UnifiedViewStrategy
+- refactor: simplify render_tasks to use unified view strategy
+- refactor: introduce UnifiedViewStrategy to compose layout and render strategies
+- refactor: create render strategy abstraction for card list rendering
+- refactor: create layout strategy abstraction for card list management
+- refactor: extract card filtering and sorting logic into card_filter_service
+- KAN-118/unfilter-tasks-list-on-completed-sprint (#93)
+- KAN-111/sprint-binding-help-is-wrong (#92)
+- KAN-33: Add Help mode with context-aware keybindings (#91)
+- ci: automatically sync develop with master after release (#90)
