+++ crates/kanban-tui/src/app.rs
@@ -11,9 +11,7 @@ use crate::{
     selection::SelectionState,
     services::{filter::CardFilter, get_sorter_for_field, BoardFilter, OrderedSorter},
     ui,
-    view_strategy::{
-        FlatViewStrategy, GroupedViewStrategy, KanbanViewStrategy, ViewRefreshContext, ViewStrategy,
-    },
+    view_strategy::{UnifiedViewStrategy, ViewRefreshContext, ViewStrategy},
 };
 use crossterm::{
     execute,
@@ -193,7 +191,7 @@ impl App {
                     ])
                     .with_multi_select(false),
             ),
-            view_strategy: Box::new(GroupedViewStrategy::new()),
+            view_strategy: Box::new(UnifiedViewStrategy::grouped()),
             card_list_component: CardListComponent::new(
                 CardListId::All,
                 CardListComponentConfig::new(),
@@ -617,9 +615,11 @@ impl App {
 
     pub fn switch_view_strategy(&mut self, task_list_view: kanban_domain::TaskListView) {
         let new_strategy: Box<dyn ViewStrategy> = match task_list_view {
-            kanban_domain::TaskListView::Flat => Box::new(FlatViewStrategy::new()),
-            kanban_domain::TaskListView::GroupedByColumn => Box::new(GroupedViewStrategy::new()),
-            kanban_domain::TaskListView::ColumnView => Box::new(KanbanViewStrategy::new()),
+            kanban_domain::TaskListView::Flat => Box::new(UnifiedViewStrategy::flat()),
+            kanban_domain::TaskListView::GroupedByColumn => {
+                Box::new(UnifiedViewStrategy::grouped())
+            }
+            kanban_domain::TaskListView::ColumnView => Box::new(UnifiedViewStrategy::kanban()),
         };
 
         self.view_strategy = new_strategy;
