+++ crates/kanban-tui/src/lib.rs
@@ -12,7 +12,9 @@ pub mod filters;
 pub mod handlers;
 pub mod input;
 pub mod keybindings;
+pub mod layout_strategy;
 pub mod markdown_renderer;
+pub mod render_strategy;
 pub mod search;
 pub mod selection;
 pub mod services;
