+++ crates/kanban-tui/src/services/mod.rs
@@ -1,5 +1,7 @@
+pub mod card_filter_service;
 pub mod filter;
 pub mod sort;
 
+pub use card_filter_service::*;
 pub use filter::*;
 pub use sort::*;
